+++ lib/ui/widgets/LinkTextEx.dart
@@ -1,14 +1,15 @@
+import 'dart:math';
+
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 import 'package:url_launcher/url_launcher.dart';
-import 'package:neom/service/Content.dart';
 
 // Copied from https://pub.dev/packages/link_text
 
 /// Easy to use text widget, which converts inlined urls into clickable links.
 /// This version checks for a TLD from a Set of known TLDs.
-class CustomLinkText extends StatefulWidget {
+class LinkTextEx extends StatefulWidget {
   /// Text, which may contain inlined urls.
   final String text;
 
@@ -30,7 +31,7 @@ class CustomLinkText extends StatefulWidget {
   /// Provides the url that was tapped.
   final void Function(String url)? onLinkTap;
 
-  const CustomLinkText(
+  const LinkTextEx(
       this.text, {
         Key? key,
         this.textStyle,
@@ -41,10 +42,13 @@ class CustomLinkText extends StatefulWidget {
       }) : super(key: key);
 
   @override
-  State<CustomLinkText> createState() => _CustomLinkTextState();
+  State<LinkTextEx> createState() => _LinkTextExState();
 }
 
-class _CustomLinkTextState extends State<CustomLinkText> {
+class _LinkTextExState extends State<LinkTextEx> {
+
+  /// Url regular expression, credits to: https://stackoverflow.com/a/63022807/3759472
+  final RegExp _urlRegex = RegExp(r"([\w+]+\:\/\/)?([\w\d-]+\.)*[\w-]+[\.\:]\w+([\/\?\=\&\#\.]?[\w-]+)*\/?");
 
   /// We hold on to recognizers so we can dispose them properly.
   final Map<String, TapGestureRecognizer> _gestureRecognizers = {};
@@ -53,15 +57,15 @@ class _CustomLinkTextState extends State<CustomLinkText> {
 
   @override
   void initState() {
-    _initTextSpans();
+    _buildTextSpans();
     super.initState();
   }
 
   @override
-  void didUpdateWidget(CustomLinkText oldWidget) {
+  void didUpdateWidget(LinkTextEx oldWidget) {
     super.didUpdateWidget(oldWidget);
     if (widget.text != oldWidget.text) {
-      _initTextSpans();
+      _buildTextSpans();
     }
   }
 
@@ -71,68 +75,49 @@ class _CustomLinkTextState extends State<CustomLinkText> {
     super.dispose();
   }
 
-  void _initTextSpans() {
+  void _buildTextSpans() {
     _textSpans.clear();
     _disposeRecognizers();
 
-    // Split text on whitespace
-    final words = widget.text.split(RegExp(r' '));
-
-    for (int i = 0; i < words.length; i++) {
-      final word = words[i];
-      if (_isPotentialLink(word)) {
-        String displayLink = word;
-        if (widget.shouldTrimParams) {
-          final questionMarkIndex = displayLink.indexOf('?');
-          if (questionMarkIndex != -1) {
-            displayLink = displayLink.substring(0, questionMarkIndex);
-          }
+    Iterable<RegExpMatch> urls = _urlRegex.allMatches(widget.text);
+    if (urls.isEmpty) {
+      // Entire text is Normal
+      _textSpans.add(TextSpan(text: widget.text, style: widget.textStyle,),);
+    }
+    else {
+      int textPos = 0;
+      for (RegExpMatch urlMatch in urls) {
+        if (textPos < urlMatch.start) {
+          // Normal word
+          String word = widget.text.substring(textPos, urlMatch.start);
+          _textSpans.add(TextSpan(text: word, style: widget.textStyle,),);
         }
-
-        TapGestureRecognizer recognizer = (_gestureRecognizers[word] ??= (TapGestureRecognizer()..onTap = () => _launchUrl(word)));
-
-        _textSpans.add(
-          TextSpan(
-            text: displayLink,
-            style: widget.linkStyle,
-            recognizer: recognizer,
-          ),
-        );
-      } else {
-        // Normal text
-        _textSpans.add(
-          TextSpan(
-            text: word,
-            style: widget.textStyle,
-          ),
-        );
+        if (urlMatch.start < urlMatch.end) {
+          // URL link
+          String url = widget.text.substring(urlMatch.start, urlMatch.end);
+          String displayUrl = url;
+          if (widget.shouldTrimParams) {
+            int pos1 = url.indexOf('?'), pos2 = url.indexOf('#');
+            int pos = (0 < pos1) ? ((0 < pos2) ? min(pos1, pos2) : pos1 ) : pos2;
+            if (0 < pos) {
+              displayUrl = url.substring(0, pos);
+            }
+          }
+          TapGestureRecognizer recognizer = (_gestureRecognizers[url] ??= (TapGestureRecognizer()..onTap = () => _launchUrl(url)));
+          _textSpans.add(TextSpan(text: displayUrl, style: widget.linkStyle, recognizer: recognizer,));
+        }
+        textPos = urlMatch.end;
       }
-
-      // Add a space after each word unless it's the last one.
-      if (i < words.length - 1) {
-        _textSpans.add(const TextSpan(text: ' '));
+      if (textPos < widget.text.length) {
+        String word = widget.text.substring(textPos);
+        _textSpans.add(TextSpan(text: word, style: widget.textStyle,),);
       }
     }
   }
 
   @override
-  Widget build(BuildContext context) {
-    return Text.rich(
-      TextSpan(children: _textSpans),
-      textAlign: widget.textAlign,
-    );
-  }
-
-  /// Checks if a 'word' might be a link by verifying it ends with a known TLD.
-  bool _isPotentialLink(String word) {
-    final lastDotIndex = word.lastIndexOf('.');
-    // Checks if dot doesn't exist or is first char
-    if (lastDotIndex < 1 || lastDotIndex == word.length - 1) {
-      return false;
-    }
-    final tldCandidate = word.substring(lastDotIndex + 1).toLowerCase();
-    return Content().topLevelDomains.contains(tldCandidate);
-  }
+  Widget build(BuildContext context) =>
+    Text.rich(TextSpan(children: _textSpans), textAlign: widget.textAlign,);
 
   void _disposeRecognizers() {
     for (final recognizer in _gestureRecognizers.values) {